記述箇所の変更＆いらないファイルの削除
